bug修改

+++ ocean-common/ocean-common-base/ocean-general-base/src/main/java/com/fantechs/common/base/general/entity/basic/history/BaseHtInAndOutRule.java
@@ -56,7 +56,7 @@ public class BaseHtInAndOutRule extends ValidGroup implements Serializable {
     @ApiModelProperty(name="warehouseId",value = "仓库ID")
     @Excel(name = "仓库ID", height = 20, width = 30,orderNum="") 
     @Column(name = "warehouse_id")
-    private Byte warehouseId;
+    private Long warehouseId;
 
     /**
      * 类型(1-入库 2-出库)
